Menu Updates

+++ includes/struct.h
@@ -6,7 +6,7 @@
 /*   By: kwiessle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/09 11:43:16 by kwiessle          #+#    #+#             */
-/*   Updated: 2016/05/12 20:19:28 by kwiessle         ###   ########.fr       */
+/*   Updated: 2016/05/17 13:16:21 by kwiessle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ typedef struct		s_param
 {
 	char			*name;
 	double			zoom;
-	int				iter;
+	long			iter;
 	int				color;
 	double			x1;
 	double			y1;
